Adjust exports so that dataAnalysis can run without attaching lrd to search path
Among adaptations necessary to get R to look for our versions of sandwich &
friends where necessary (ie for lmrobs) while utilizing the ordinary sandwich package
versions where it can, this one promotes safety, i.e. may avoid some bugs
due to inadvertently referencing the wrong version.

Our `testSH()` appears to be set up in such a way that if we had inadvertently
been referencing the wrong sandwich version, the calculations would continue with a
non-sandwich SE, quietly.  I don't know whether this was happening or not, but if it
was then it may have produced some results that were in error, by the our standards.
These errors would be corrected in this commit.  If we later find ourselves
trying to track down vexing discrepancies in results, here is a good place to look.