Fix incomplete merge of fullOutcomeSim.Rmd in [master 9f6785f]
From reading the fullOutcomeSim.md that was committed in [master 4fa3b4a],
it's the first half of the unmerged file that was used to produce those
computations.  So that's the half that stays in the source file after this
commit.  (The deleted version appears to differ from the version of the code that
was retained only in its markdown comments, not the code itself.)