Post-hoc cosmetic update of simulation script
1. There was a bug affecting printing of "full" polynomial simulation results,
   squashed in [master 8ed363cc]. This commit grafts in output as produced w/o bug.
2.  While I was at it, made a few more touch-ups fullOutcomeSim.Rmd, in [master cb350763].
   These are folded in here two.

To make this change, I ran fullOutcomeSim.Rmd with nreps set to 0, so that it loaded in
the saved simulation artifacts rather than regenerating them.  What's committed here is
a partial merge of that md file with the one originally generated at the sime that the
simulations were actually run.