Shorten inital disc of robust regr, commenting out mystery of fall 2014
Cf. #30 . With all we're trying to get out in the intro, the fall 2014
death spike seems an unwelcome distraction.
Perhaps there should be a self-standing section presenting
reasons for use of robust regression.  In that context the mystery
of fall 2014 is a nicely tangible complement to the more abstract
point about the LSO sample potentially being somewhat contaminated
even after the donut maneuver.  Perhaps these issues can be postponed
until after intro and discussion of the "main method".