Modified line about entropy
Dan M didn't like our use of the word entropy; I (AS) changed it to "randomness." I know that doesn't have the same technical meaning, but it will throw fewer people off, and it works nicely with the phrase "basis for inference." And it keeps the main point intact, I think. Also, we no longer discuss conditioning on (Y_C,n_t,n_c) so I replaced that with "conditioning on unadjusted Y_C".