Simplify, rationalize sec 2 title
Not all ofthe methods being reviewed in this section are RDD methods.

At this point in the paper we have yet to explain to the reader why we think he should care about robustness.

So, I think we just have to ask him to trust us to select out what needs reviewing in order to set the stage for what's new.

For that it helps to signal clearly that this section is about review.