Adjust wording to remove a confusing & misleading passage
The removes the sentence
> Following this ``limit understanding,'' it is common to
> analyze RDDs using regression.
which seems to make the odd claim that the limit understanding
tells you to use (ordinary, global) regression.  This seems to be
a misplaced edit of this sentence appearing in earlier versions,
which is now restored:
> Following this ``limit understanding,'' it is common to analyze RDDs
> using regression to estimate the functional relationships of $r$
> to $\EE( Y | R=r)$ on either side of the cutoff.
Probably that sentence was split in order to introduce a definition
of $Y$.  In this commit that need is addressed by introducing that
symbol within the previous paragraph.