Address readibility of abstract
The last formulation, contrasting "the limits of regressions on the running
variable... as it approaches", was a little bit off in terms of
both syntax and use of metaphor.  This version does without terms of
art, even to the point of changing the topic from RD "designs" to
RD "studies" -- but I think we don't need to burden the reader with
all that at the stage of the abstract.